adding some stuff

+++ src/modules/users/services/RemoveUserService.ts
@@ -1,3 +1,4 @@
+import AppError from '../../../errors/AppError';
 import UsersRepository from '../repositories/UsersRepository';
 
 interface IRequest {
@@ -7,6 +8,12 @@ interface IRequest {
 class RemoveUserService {
   public async execute({ id }: IRequest): Promise<void> {
     const usersRepository = new UsersRepository();
+
+    const checkUserExists = await usersRepository.findById(id);
+
+    if (!checkUserExists) {
+      throw new AppError('User does not exists');
+    }
 
     await usersRepository.deleteUser(id);
   }
